- Refactored primefaces selenium adapter to be a base class extended by chrome/firefox variants and   specified via a system property to a filtered config.properties maven test resource - Added maven properties to specify primefaces selenium adapter implementation and chrome driver binary   along with a maven profile to allow overriding the adapter implementation property for chrome - Added chrome job to travis - Updated README.md to add chrome notes
