+++ src/test/resources-filtered/primefaces-selenium/config.properties
@@ -1,0 +1,1 @@
+adapter=${primefaces-selenium.adapter}
