--- src/test/resources/primefaces-selenium/config.properties
@@ -1,1 +1,0 @@
-adapter=org.primefaces.extensions.integrationtests.PrimeFacesSeleniumAdapterImpl
